Introduce sampler interface
