+++ sbi/analysis/__init__.py
@@ -1,6 +1,9 @@
 from sbi.analysis.conditional_density import (
     conditional_corrcoeff,
     eval_conditional_density,
+    parameter_conditonal_potential,
+    parameter_conditional_mdn,
 )
 from sbi.analysis.plot import conditional_pairplot, pairplot
 from sbi.analysis.sensitivity_analysis import ActiveSubspace
+from sbi.analysis.get_maximum import get_maximum
