+++ sbi/inference/snpe/__init__.py
@@ -1,4 +1,4 @@
-from sbi.inference.snpe.snpe_a import SNPE_A  # noqa: F401
-from sbi.inference.snpe.snpe_b import SNPE_B  # noqa: F401
+# from sbi.inference.snpe.snpe_a import SNPE_A  # noqa: F401
+# from sbi.inference.snpe.snpe_b import SNPE_B  # noqa: F401
 from sbi.inference.snpe.snpe_base import PosteriorEstimator  # noqa: F401
 from sbi.inference.snpe.snpe_c import SNPE_C  # noqa: F401
