--- sbi/mcmc/__init__.py
@@ -1,4 +1,0 @@
-from sbi.mcmc.init_strategy import IterateParameters, prior_init, sir
-from sbi.mcmc.slice import Slice
-from sbi.mcmc.slice_numpy import SliceSampler
-from sbi.mcmc.slice_numpy_vectorized import SliceSamplerVectorized
