+++ sbi/samplers/__init__.py
@@ -1,0 +1,3 @@
+from sbi.samplers.mcmc.slice import Slice
+from sbi.samplers.mcmc.slice_numpy import SliceSampler
+from sbi.samplers.mcmc.slice_numpy_vectorized import SliceSamplerVectorized
