+++ sbi/samplers/mcmc/__init__.py
@@ -1,0 +1,4 @@
+from sbi.samplers.mcmc.init_strategy import IterateParameters, prior_init, sir
+from sbi.samplers.mcmc.slice import Slice
+from sbi.samplers.mcmc.slice_numpy import SliceSampler
+from sbi.samplers.mcmc.slice_numpy_vectorized import SliceSamplerVectorized
