+++ sbi/samplers/mcmc/slice_numpy.py
