+++ sbi/utils/__init__.py
@@ -3,7 +3,6 @@ from sbi.utils.conditional_density import (
     conditional_corrcoeff,
     eval_conditional_density,
     extract_and_transform_mog,
-    condition_mog,
 )
 from sbi.utils.get_nn_models import classifier_nn, likelihood_nn, posterior_nn
 from sbi.utils.io import get_data_root, get_log_root, get_project_root
@@ -24,9 +23,6 @@ from sbi.utils.sbiutils import (
     mask_sims_from_prior,
     mcmc_transform,
     mog_log_prob,
-    optimize_potential_fn,
-    rejection_sample,
-    rejection_sample_posterior_within_prior,
     standardizing_net,
     standardizing_transform,
     warn_if_zscoring_changes_data,
@@ -34,6 +30,7 @@ from sbi.utils.sbiutils import (
     warn_on_invalid_x_for_snpec_leakage,
     within_support,
     x_shape_from_simulation,
+    match_theta_and_x_batch_shapes,
 )
 from sbi.utils.torchutils import (
     BoxUniform,
@@ -69,3 +66,4 @@ from sbi.utils.user_input_checks import (
     validate_theta_and_x,
 )
 from sbi.utils.user_input_checks_utils import MultipleIndependent
+from sbi.utils.potentialutils import transformed_potential, pyro_potential_wrapper
