+++ sbi/utils/conditional_density.py
@@ -1,16 +1,20 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-from typing import Any, Callable, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from warnings import warn
+from copy import deepcopy
+from functools import partial
 
+import torch.distributions.transforms as torch_tf
 import torch
-from torch import Tensor
-
+from torch import Tensor, nn
+import numpy as np
+from sbi.types import Shape
 from pyknos.mdn.mdn import MultivariateGaussianMDN as mdn
 from pyknos.nflows.flows import Flow
 from sbi.utils.torchutils import ensure_theta_batched
-from sbi.utils.torchutils import BoxUniform
+from sbi.utils.torchutils import BoxUniform, ScalarFloat, atleast_2d_float32_tensor
 
 
 def eval_conditional_density(
@@ -462,3 +466,122 @@ def condition_mog(
 
     sumlogdiag = torch.sum(torch.log(torch.diagonal(precfs_xx, dim1=2, dim2=3)), dim=2)
     return logits, means, precfs_xx, sumlogdiag
+
+
+def build_conditioned_potential_fn(
+    potential_fn: Callable, condition: Tensor, dims_to_sample: List[int]
+) -> Callable:
+    def conditioned_potential_fn(
+        theta: np.ndarray, track_gradients: bool = False
+    ) -> ScalarFloat:
+        r"""
+        Return conditional posterior log-probability or $-\infty$ if outside prior.
+
+        The only differences to the `np_potential` is that it tracks the gradients and
+        does not return a `numpy` array.
+
+        Args:
+            theta: Free parameters $\theta_i$, batch dimension 1.
+
+        Returns:
+            Conditional posterior log-probability $\log(p(\theta_i|\theta_j, x))$,
+            masked outside of prior.
+        """
+        theta_ = ensure_theta_batched(torch.as_tensor(theta, dtype=torch.float32))
+        theta_condition = deepcopy(condition)
+        theta_condition = theta_condition.repeat(theta_.shape[0], 1)
+        theta_condition[:, dims_to_sample] = theta_
+
+        return potential_fn(theta_condition, track_gradients=track_gradients)
+
+    return conditioned_potential_fn
+
+
+class RestrictedPriorForConditional:
+    """
+    Class to restrict a prior to fewer dimensions as needed for conditional sampling.
+
+    The resulting prior samples only from the free dimensions of the conditional.
+
+    This is needed for the the MCMC initialization functions when conditioning.
+    For the prior init, we could post-hoc select the relevant dimensions. But
+    for SIR, we want to evaluate the `potential_fn` of the conditional
+    posterior, which takes only a subset of the full parameter vector theta
+    (only the `dims_to_sample`). This subset is provided by `.sample()` from
+    this class.
+    """
+
+    def __init__(self, full_prior: Any, dims_to_sample: List[int]):
+        self.full_prior = full_prior
+        self.dims_to_sample = dims_to_sample
+
+    def sample(self, *args, **kwargs):
+        """
+        Sample only from the relevant dimension. Other dimensions are filled in
+        by the `ConditionalPotentialFunctionProvider()` during MCMC.
+        """
+        return self.full_prior.sample(*args, **kwargs)[:, self.dims_to_sample]
+
+    def log_prob(self, *args, **kwargs):
+        r"""
+        `log_prob` is same as for the full prior, because we usually evaluate
+        the $\theta$ under the full joint once we have added the condition.
+        """
+        return self.full_prior.log_prob(*args, **kwargs)
+
+
+class RestrictedTransformForConditional(nn.Module):
+    """
+    Class to restrict the transform to fewer dimensions for conditional sampling.
+
+    The resulting transform transforms only the free dimensions of the conditional.
+    Notably, the `log_abs_det` is computed given all dimensions. However, the
+    `log_abs_det` stemming from the fixed dimensions is a constant and drops out during
+    MCMC.
+
+    This is needed for the the MCMC initialization functions when conditioning and
+    when transforming the samples back into the original theta space after sampling.
+    """
+
+    def __init__(
+        self,
+        transform: torch_tf.Transform,
+        condition: Tensor,
+        dims_to_sample: List[int],
+    ) -> None:
+        super().__init__()
+        self.transform = transform
+        self.condition = ensure_theta_batched(condition)
+        self.dims_to_sample = dims_to_sample
+
+    def forward(self, theta: Tensor) -> Tensor:
+        r"""
+        Transform restricted $\theta$.
+        """
+        full_theta = self.condition.repeat(theta.shape[0], 1)
+        full_theta[:, self.dims_to_sample] = theta
+        tf_full_theta = self.transform(full_theta)
+        return tf_full_theta[:, self.dims_to_sample]
+
+    def inv(self, theta: Tensor) -> Tensor:
+        r"""
+        Inverse transform restricted $\theta$.
+        """
+        full_theta = self.condition.repeat(theta.shape[0], 1)
+        full_theta[:, self.dims_to_sample] = theta
+        tf_full_theta = self.transform.inv(full_theta)
+        return tf_full_theta[:, self.dims_to_sample]
+
+    def log_abs_det_jacobian(self, theta1: Tensor, theta2: Tensor) -> Tensor:
+        """
+        Return the `log_abs_det_jacobian` of |dtheta1 / dtheta2|.
+
+        The determinant is summed over all dimensions, not just the `dims_to_sample`
+        ones.
+        """
+        full_theta1 = self.condition.repeat(theta1.shape[0], 1)
+        full_theta1[:, self.dims_to_sample] = theta1
+        full_theta2 = self.condition.repeat(theta2.shape[0], 1)
+        full_theta2[:, self.dims_to_sample] = theta2
+        log_abs_det = self.transform.log_abs_det_jacobian(full_theta1, full_theta2)
+        return log_abs_det
